// src/app/shared/models/proyeccionCostos/proyeccionCostos.model.ts
export class ProyeccionCostosModel{
    concepto: string;
    clasificacion: string;
    real: number;
    realP: number;
    presupuesto: any;
    presupuestoP: number;
    presupuestoPC: number;
    proyeccion: number;
    proyeccionP: number;
    proyeccionPC: number;
    proyeccionFactor: number;
    periodo: number;
    renglon: number;
}